requirements.txt
# Основные зависимости для утилит и скриптов обслуживания
# после миграции на Edge Functions
supabase
python-dotenv
google-genai